mpm: gpsd_iface: Make GPGGA generation more robust
The GPGGA string generation relies on TPV and SKY messages being
accurate, meaning they are a list of dictionaries. There have been cases
where this was inaccurate (an empty list was returned). MPM would show
errors as such:

[ERROR] [MPM.RPCServer] Uncaught exception in method get_mb_sensor :list
index out of range
 Traceback (most recent call last):
   File "/usr/lib/python3.7/site-packages/usrp_mpm/rpc_server.py", line
   184, in new_claimed_function
       return function(*args)
         File
	 "/usr/lib/python3.7/site-packages/usrp_mpm/periph_manager/base.py",
	 line 1000, in get_mb_sensor
	     self, self.mboard_sensor_callback_map.get(sensor_name)
	       File
	       "/usr/lib/python3.7/site-packages/usrp_mpm/gpsd_iface.py",
	       line 313, in get_gps_gpgga_sensor
	           tpv_sensor_data = gps_info.get('tpv', [{}])[0]
		   IndexError: list index out of range

This is a patch to check that the lists are not empty before directly
referencing index 0 therein.