multi_usrp: Factor out make_overall_tune_range() and fix limits
This function had an issue where it might return negative frequency
values. A quick fix was to limit it to positive frequencies.

Since this function was duplicated between multi_usrp and
multi_usrp_rfnoc, this patch also moves it to a common location to not
have to fix it twice.