docs: Fix typo in ZBX Block Diagram
This commit changes the ZBX block diagram which is also visible at:

  https://files.ettus.com/manual/page_zbx.html

The old version of this file had a typo where two paths in the RF
section of the RX chain were both labeled "RF 3". The coresponding path
in the TX chain was labeled "RF 4".

Rather than change the RX label to "RF 4", this commit changes both
labels to "RF 0", since that is how they are referred to in the tuning
tables in the UHD source (see zbx_constants.hpp). This naming convention
makes more sense anyways, because this path bypasses the filter bank in
the RF section.

Signed-off-by: Sam O'Brien <sam.obrien@ni.com>