utils: Get signal above noise floor when finding optimal gain
Ensure that the signal is at least visible above the noise floor before
attempting to identify gain compression. Another option that was
considered involved starting from the top of the gain range and working
down to avoid this issue, however, that option could expose the device
to unsafe amounts of incoming signal power.

Signed-off-by: mattprost <matt.prost@ni.com>