ci: Add custom boost version support
This adds new variables which can be used to set a a URL to a custom
Boost archive. If provided, the pipeline will download the custom Boost
archive, unpack it, build it, and provide CMake flags for UHD before
starting the UHD build itself.

This allows semi-automatic testing of new Boost versions as they are
released.