python: rfnoc: Change reference type for noc_block_base export
By changing the type for accesses to noc_block_base calls in the Python
from sptr& to a simple reference (&), we fix the "holder type" issues
that crop up when trying to use radio_control from multi_usrp, which
returns access to the block as a reference rather than a `sptr`.

The error message seen without this fix always contains this string:

    Unable to cast from non-held to held instance (T& to Holder<T>)

(The exact message depends on the API call made).