cmake: tests: Add build-python path to PYTHONPATH
This is more of an expressive change than a functional change; Python
seems to add this path to the PYTHONPATH anyway, at least for some
systems. We neverless make this change because:
- It's more explicit/expressive. When tests are run, the PYTHONPATH env
  variable is printed, and it now contains this path where it should be,
  right at the front. People reading the ctest/python.unittest output
  now get told explicitly which path we mean.
- This guarantees that this path is added, even if Python/unittest
  should behave differently on other systems or versions.

To clarify: When running unit tests, we want to run the Python code from
build/python, not the installed version. The latter may not yet exist,
and if it does, it's not the version we are editing.