dissectors: Fix inclusion of glib.h and Python version
Previously, we were doing this:

```cpp
extern "C" {
}
```

This is not how glib.h is supposed to be included, according to their
documentation. It turns out that it happened to work for a long time,
and lots of projects include glib.h that way, and the glib devs are
trying to accommodate for those cases. How nice of them!

However, on Fedora 34, we have a version of glib which does not have
a workaround patch, so we include glib.h the it should be.

The second issue was that a build script required the existence of
a `python` executable, which is not always available. A `python3`
executable can be assumed, since that's a dependency for UHD, too.