B200: Re-sync times
The times on the device can glitch if either the tick rate changes or
the number of active chains changes.  This throws off the time if the
user gets streamers, changes the sample rate, or changes the tick rate
after synchronizing the time.  This change re-synchronizes the times
automatically in those cases.

Signed-off-by: michael-west <michael.west@ettus.com>