rfnoc: Fix issue in uhd::rfnoc::connect_through_blocks()
When connect_through_blocks() was called on blocks within a single
chain, there was a bug where the chain was incorrectly cropped. In
a standard FPGA image, say one was to use this API call to connect the
radio to the DDC. It would generate a chain of blocks hanging off the
radio as such:

Radio -> DDC -> SEP

What the code should do, and what this fix provides, is that the chain
gets cropped after the DDC, to look like this:

Radio -> DDC

With the current bug, it would assume the chain has a dangling edge, and
incorrectly throw an exception.
Note that this bug would not appear when source and destination block
are on separate chains (i.e., both have an SEP in their chain).

This patch includes minor logging and comment improvements around the
offending lines of code.