rfnoc: replay block: Disable prop and action propagation
The replay block is more like the radio block than like a FIFO. In
particular, consider this flow graph:

    Replay -> DDC -> Replay

Imagine you're using the replay block to test the DDC block with
prerecorded data. If we treated the Replay Block like a FIFO, then we'd
have a loop in the graph (which is already wrong). If we used the DDC to
resample, then the input- and output sample rate of the Replay mismatch,
which is a legal way to use the Replay block, but not possible if we
treat the graph like a loop.