e31x/e320: Amend LO-locked sensor names
In UHD 3, we had two sensors names for LO lock on these devices:
lo_lock, and lo_locked. The latter is the more standard, and is checked
in examples like rx_samples_to_file.

In UHD 4, the latter was removed without comment. This adds the sensor
back again and also updates the documentation accordingly.