tests: Add replay-back-edge test
This tests the following graph: DDC -> Replay -> DDC, where the initial
and final blocks are the same (i.e., a loop). This could be useful for
generating and capturing data with the same replay block while testing
a block.

Note that this test will fail if the edge consistency checks are buggy.