RFNoC: Fix DSP frequency accuracy
The host code was calculating and programming a 32-bit value for the DSP
frequency, but the DDS modules in the FPGA only use the upper 24-bits.
This led to inaccurate frequency values being returned.  This change
corrects the resolution of the value on the host side so an accurate
value is returned.

Signed-off-by: michael-west <michael.west@ettus.com>