docs: Fix MathJax formulae rendering
When a webpage is accessed via secure HTTP, and that webpage attempts to
address active content via a non-secure URI, most modern browsers will
block the loading of that content as a security precaution. In this
case, the URI to the MathJax JavaScript rendering library was specified
in the Doxygen configuration with an HTTP (i.e., non-encrypted) URI,
thus preventing the browser from loading it and rendering formulae
correctly.