rfnoc: transport: Check if streamers are connected in send() and recv()
This adds a check in send() and recv() whether or not the streamer is
actually connected. If not, an exception is thrown with the message:

    [rx_stream] Attempting to call recv() before all channels are connected!

or

    [tx_stream] Attempting to call send() before all channels are connected!

The check is a single boolean flag check, but it does add a branch in
our hot code. Since this event is unlikely, and only happens in badly
configured apps, we will get some help from the CPUs branch prediction
to reduce the additional cost of this check.